+++ Maxim_Khromov/homework5/views/stopwatch/stopwatch.hbs
@@ -17,4 +17,4 @@
         {{/each}}
     </div>
 </article>
-<script src="components/stopwatch/stopwatch.js"></script>
+<script src="/stopwatch/stopwatch.js"></script>
